Gateway/requirements.txt
fastapi[standard]~=0.115.12
pydantic-settings
loguru~=0.7.3
python-dotenv~=1.1.0
pyjwt~=2.10.1
passlib[bcrypt]~=1.7.4
starlette~=0.46.2
pydantic~=2.11.3
sqlalchemy~=2.0.41
psycopg[binary]
alembic
aiosmtplib~=4.0.1
aiosmtpd~=1.4.6
redis>=4.3.0
asyncpg~=0.30.0